pyproject.toml
[tool.poetry]
name = "constgen"
version = "0.2.0"
description = "ConstGen lets you share multiple object definitions between multiple programming languages"
authors = ["Jordan Cohen <yarden.mec@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ymec/constgen.git"
repository = "https://github.com/ymec/constgen.git"
#packages = [
#    { include = "constgen" } ,
#    { include = "constgen/config/**/*.py" },
#    { include = "constgen/generators" },
#]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^5.4.1"
pydantic = "^1.8.1"
inflection = "^0.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
constgen = 'constgen.constgen:main'
